requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
astroid==2.5.1
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cssselect==1.1.0
discord.py==1.7.3
fake-useragent==0.1.11
feedparser==6.0.2
idna==2.10
isort==5.7.0
lazy-object-proxy==1.5.2
lxml==4.6.3
mccabe==0.6.1
multidict==5.1.0
numpy==1.20.1
pandas==1.2.3
parse==1.19.0
pyee==8.1.0
pylint==2.7.2
pyppeteer==0.2.5
pyquery==1.4.3
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-html==0.10.0
selenium==3.141.0
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2.1
toml==0.10.2
tqdm==4.59.0
typing-extensions==3.10.0.0
urllib3==1.26.4
w3lib==1.22.0
websockets==8.1
wrapt==1.12.1
yahoo-fin==0.8.7
yarl==1.6.3